Address PR review feedback
Cast expiresIn to int before serializing to X-Amz-Expires
to ensure whole-second values. Remove unnecessary
#if ASYNC_AWAIT guards from new test methods. Add DevConfig
file for changelog generation.

+++ sdk/src/Services/DSQL/Custom/Util/DSQLAuthTokenGenerator.cs
@@ -406,7 +406,7 @@ namespace Amazon.DSQL.Util
 
             request.UseQueryString = true;
             request.HttpMethod = HTTPGet;
-            request.Parameters.Add(XAmzExpires, expiresIn.TotalSeconds.ToString(CultureInfo.InvariantCulture));
+            request.Parameters.Add(XAmzExpires, ((int)expiresIn.TotalSeconds).ToString(CultureInfo.InvariantCulture));
             request.Parameters.Add(ActionKey, actionValue);
             request.Endpoint = new UriBuilder(HTTPS, hostname).Uri;
